+++ inst/yupanapro/setup.r
@@ -4,7 +4,7 @@
 #> open https://flavjack.github.io/inti/
 #> open https://flavjack.shinyapps.io/yupanapro/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2020-10-30
+#> date .: 2020-11-13
 # -------------------------------------------------------------------------
 
 # -------------------------------------------------------------------------
@@ -26,7 +26,7 @@ cran <- c(
   )
 
 git <- c(
-  "rstudio/bootstraplib"
+  "rstudio/bslib"
   , "Flavjack/inti" 
   )